refactor(payments-stripe, auth): Update Coupon API endpoint
