+++ libs/payments/stripe/src/lib/promotionCode.manager.ts
@@ -4,11 +4,23 @@
 
 import { Injectable } from '@nestjs/common';
 
+import { ProductManager } from './product.manager';
 import { StripeClient } from './stripe.client';
+import { PromotionCodeCouldNotBeAttachedError } from './stripe.error';
+import {
+  checkSubscriptionPromotionCodes,
+  checkValidPromotionCode,
+  getSubscribedPrice,
+} from './stripe.util';
+import { SubscriptionManager } from './subscription.manager';
 
 @Injectable()
 export class PromotionCodeManager {
-  constructor(private client: StripeClient) {}
+  constructor(
+    private client: StripeClient,
+    private productManager: ProductManager,
+    private subscriptionManager: SubscriptionManager
+  ) {}
 
   async retrieve(id: string) {
     return this.client.promotionCodesRetrieve(id);
@@ -21,5 +33,68 @@ export class PromotionCodeManager {
     });
 
     return promotionCodes.data.at(0);
+  }
+
+  async applyPromoCodeToSubscription(
+    customerId: string,
+    subscriptionId: string,
+    promotionId: string
+  ) {
+    try {
+      const subscription = await this.subscriptionManager.retrieve(
+        subscriptionId
+      );
+      if (subscription?.status !== 'active')
+        throw new PromotionCodeCouldNotBeAttachedError(
+          'Subscription is not active',
+          undefined,
+          { subscriptionId }
+        );
+      if (subscription.customer !== customerId)
+        throw new PromotionCodeCouldNotBeAttachedError(
+          'subscription.customerId does not match passed in customerId',
+          undefined,
+          {
+            customerId,
+            subscriptionId,
+          }
+        );
+
+      const promotionCode = await this.retrieve(promotionId);
+
+      checkValidPromotionCode(promotionCode);
+
+      const price = getSubscribedPrice(subscription);
+      const productId = price.product;
+      const product = await this.productManager.retrieve(productId);
+
+      checkSubscriptionPromotionCodes(promotionCode, price, product);
+
+      const updatedSubscription = await this.subscriptionManager.update(
+        subscriptionId,
+        {
+          discounts: [
+            {
+              promotion_code: promotionId,
+            },
+          ],
+        }
+      );
+      return updatedSubscription;
+    } catch (err) {
+      if (err.type === 'StripeInvalidRequestError') {
+        throw new PromotionCodeCouldNotBeAttachedError(
+          'Promotion code could not be attached to subscription',
+          err,
+          {
+            customerId,
+            subscriptionId,
+            promotionId,
+          }
+        );
+      } else {
+        throw err;
+      }
+    }
   }
 }
